pyproject.toml
[tool.poetry]
name = "waveforge-manager"
version = "0.1.0"
description = ""
authors = ["SuperBaumanets <slavakocherg@mail.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.13"  
requests = "^2.32.3"
numpy = "^1.26.4"         
pandas = "^2.2.3"
pyside6 = "^6.8.2.1"      
matplotlib = "^3.10.0"
scipy = "^1.15.2"
pyqtgraph = "^0.13.7"
toml = "^0.10.2"
tomli = "^2.2.1"
tomli-w = "^1.2.0"
pytest = "^8.3.5"
pydantic = "^2.10.6"
pyadi-iio = "^0.0.19"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.4.2"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"